fix(ci): python-coverage-shards — split secret-metadata test out of oversize provisioner part_001
